refractor: fileGet functions -> fileGetUtil.ts
